Move settings from creator to directory plugin
